requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
transformers==4.38.2
torch==2.2.1
tokenizers==0.15.2